test(typecheck): split executor tests
